Fix dangling ref

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -1,8 +1,7 @@
 package seedu.address.logic.commands;
 
-import static java.util.Objects.requireNonNull;
-
 import java.util.List;
+import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.ToStringBuilder;
@@ -24,6 +23,9 @@ public class DeleteCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_ATTRACTION_SUCCESS = "Deleted Attraction: %1$s";
+    public static final String MESSAGE_ATTRACTION_IN_ITINERARY
+            = "Cannot delete attraction as it is referenced in one or more itineraries. "
+            + "Please remove it from all itineraries first.";
 
     private final Index targetIndex;
 
@@ -41,6 +43,11 @@ public class DeleteCommand extends Command {
         }
 
         Attraction attractionToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        if (model.isAttractionInAnyItinerary(attractionToDelete)) {
+            throw new CommandException(MESSAGE_ATTRACTION_IN_ITINERARY);
+        }
+
         model.deleteAttraction(attractionToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_ATTRACTION_SUCCESS, Messages.format(attractionToDelete)));
     }
